Make the error definitions public
